Pull up local app server to WebServiceTestCase (fixes #163)
