+++ test/service/test_request_validation.py
@@ -1,14 +1,8 @@
 import os
 import sys
 import logging
-import time
 from unittest import mock
-
 import requests
-from threading import Thread
-from chalice.local import LocalDevServer
-from chalice.config import Config as ChaliceConfig
-from azul import Config as AzulConfig
 from service import WebServiceTestCase
 
 log = logging.getLogger(__name__)
@@ -20,73 +14,11 @@ def setUpModule():
     log.addHandler(stream_handler)
 
 
-class ChaliceServerThread(Thread):
-    def __init__(self, app, config, host, port):
-        super().__init__()
-        self.server_wrapper = LocalDevServer(app, config, host, port)
-
-    def run(self):
-        self.server_wrapper.serve_forever()
-
-    def kill_thread(self):
-        self.server_wrapper.server.shutdown()
-        self.server_wrapper.server.server_close()
-
-    def address(self):
-        return self.server_wrapper.server.server_address
-
-
 class FacetNameValidationTest(WebServiceTestCase):
     filter_facet_message = {"Code": "BadRequestError",
                             "Message": "BadRequestError: Unable to filter by undefined facet bad-facet."}
     sort_facet_message = {"Code": "BadRequestError",
                           "Message": "BadRequestError: Unable to sort by undefined facet bad-facet."}
-
-    @property
-    def base_url(self):
-        address = self.server_thread.address()
-        return f"http://{address[0]}:{address[1]}/"
-
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.path_to_app = os.path.join(AzulConfig().project_root, 'lambdas', 'service')
-        sys.path.append(cls.path_to_app)
-        from app import app
-        cls.app = app
-
-    @classmethod
-    def tearDownClass(cls):
-        super().tearDownClass()
-        sys.path.remove(cls.path_to_app)
-
-    def setUp(self):
-        super().setUp()
-        log.debug("Setting up tests")
-        log.debug("Created Thread")
-        self.server_thread = ChaliceServerThread(self.app, ChaliceConfig(), 'localhost', 0)
-        log.debug("Started Thread")
-        self.server_thread.start()
-        deadline = time.time() + 10
-        while True:
-            url = self.base_url
-            try:
-                response = requests.get(url)
-                response.raise_for_status()
-            except Exception:
-                if time.time() > deadline:
-                    raise
-                log.debug("Unable to connect to server", exc_info=True)
-                time.sleep(1)
-            else:
-                break
-
-    def tearDown(self):
-        log.debug("Tearing Down Data")
-        self.server_thread.kill_thread()
-        self.server_thread.join(timeout=10)
-        if self.server_thread.is_alive():
-            self.fail('Thread is still alive after joining')
 
     def test_hello_world(self):
         url = self.base_url
